requirements.txt
attrs==20.3.0
aws-cdk.assets==1.104.0
aws-cdk.aws-cloudwatch==1.104.0
aws-cdk.aws-ec2==1.104.0
aws-cdk.aws-efs==1.104.0
aws-cdk.aws-events==1.104.0
aws-cdk.aws-iam==1.104.0
aws-cdk.aws-kms==1.104.0
aws-cdk.aws-logs==1.104.0
aws-cdk.aws-s3==1.104.0
aws-cdk.aws-s3-assets==1.104.0
aws-cdk.aws-ssm==1.104.0
aws-cdk.cloud-assembly-schema==1.104.0
aws-cdk.core==1.104.0
aws-cdk.cx-api==1.104.0
aws-cdk.region-info==1.104.0
boto3==1.17.74
botocore==1.20.74
cattrs==1.5.0
constructs==3.3.75
jmespath==0.10.0
jsii==1.29.0
publication==0.0.3
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
